requirements.txt
slack-bolt>=1.18.0
pyyaml>=6.0.1
aiohttp>=3.9.1
python-dotenv>=1.0.0
pydantic>=2.6.1
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0
types-PyYAML>=6.0.12.12